NPCSystem - Metahuman
New folder which includes the NPC system for metahumans, currently it uses a mannequin character. I have started to implement a interface so BP_Spline and the FollowSplineComponent can communicate when to stop walking so other animations can be played out.

The TestLevel map has been update to include tests for the npc system.