Throw at npc
Currently when hit with "ThrowStone" blueprint the npc will stop for 1 second then continue on. I am planning on finding some animations of a stunned state that will be used when hit, I also need to figure out the velocity of the object so I know when to ignore the hit.